\name{egDataset1}
\Rdversion{1.1}
\alias{egDataset1}
\docType{data}
\title{
 Example dataset used in reference paper.
}
\description{
  Consists of 3 sets of 25 points, each from a multivariate distribution 
  with differing means.
}
\usage{data(egDataset1)}
\references{
  Fuentes, C. and Casella, G. (2009) Testing for the Existence of Clusters. 
  \emph{SORT} \bold{33}(2)
}
\keyword{datasets}
